Milvus/requirements.txt
pymilvus==2.3.3
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
pillow>=8.3.0
